SharedPreference 작업 공간 이동
