+++ app/src/main/java/campus/tech/kakao/map/viewmodel/MainViewModel.kt
@@ -80,4 +80,8 @@ class MainViewModel(
 
         updateTabViewVisible()
     }
+
+    fun saveLastLocation() {
+        placeRepository.saveLastLocation()
+    }
 }
